Ucheba prepodavatelyam makety

+++ app/Http/Controllers/Admin/UchebaPrepodavatelyamMaketyController.php
@@ -7,18 +7,18 @@ use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Storage;
-use App\Models\PedagogicheskijSostavDokumenty;
+use App\Models\UchebaPrepodavatelyamMakety;
 
-class PedagogicheskijSostavDokumentyController extends Controller
+class UchebaPrepodavatelyamMaketyController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index(): View
     {
-        $documents = PedagogicheskijSostavDokumenty::orderBy('id', 'desc')->get();
+        $documents = UchebaPrepodavatelyamMakety::orderBy('id', 'desc')->get();
 
-        return view('dashboard.pedagogicheskij-sostav-dokumenty', compact('documents'));
+        return view('dashboard.ucheba-prepodavatelyam-makety', compact('documents'));
     }
 
     /**
@@ -26,13 +26,13 @@ class PedagogicheskijSostavDokumentyController extends Controller
      */
     public function create(): View
     {
-        return view('dashboard.pedagogicheskij-sostav-dokumenty-create');
+        return view('dashboard.ucheba-prepodavatelyam-makety-create');
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request): RedirectResponse
     {
         $validated = $request->validate([
             'title' => 'required|min:4|max:250',
@@ -48,16 +48,16 @@ class PedagogicheskijSostavDokumentyController extends Controller
         ]);
 
         // Файл
-        $file = (new \App\Services\FileAdd($validated['input-main-file'], 'pedagogicheskij-sostav-dokumenty'))->add();
+        $file = (new \App\Services\FileAdd($validated['input-main-file'], 'prepodavatelyam-makety'))->add();
 
         // Подпись
-        $sig_file = array_key_exists('input-sig-file', $validated) ? (new \App\Services\FileAdd($validated['input-sig-file'], 'pedagogicheskij-sostav-dokumenty'))->add() : NULL;
+        $sig_file = array_key_exists('input-sig-file', $validated) ? (new \App\Services\FileAdd($validated['input-sig-file'], 'prepodavatelyam-makety'))->add() : NULL;
         
         // Ключ
-        $key_file = array_key_exists('input-key-file', $validated) ? (new \App\Services\FileAdd($validated['input-key-file'], 'pedagogicheskij-sostav-dokumenty'))->add() : NULL;
+        $key_file = array_key_exists('input-key-file', $validated) ? (new \App\Services\FileAdd($validated['input-key-file'], 'prepodavatelyam-makety'))->add() : NULL;
 
         // Создание модели
-        PedagogicheskijSostavDokumenty::create([
+        UchebaPrepodavatelyamMakety::create([
             'title' => $validated["title"],
             'file' => $file,
             'sig_file' => $sig_file,
@@ -65,7 +65,7 @@ class PedagogicheskijSostavDokumentyController extends Controller
             'filetype' => $validated["input-main-file"]->getClientOriginalExtension()
         ]);
 
-        return redirect('/dashboard/pedagogicheskij-sostav-dokumenty');
+        return redirect('/dashboard/ucheba-prepodavatelyam-makety');
     }
 
     /**
@@ -79,17 +79,17 @@ class PedagogicheskijSostavDokumentyController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(string $id): View
     {
-        $document = PedagogicheskijSostavDokumenty::findOrFail($id);
+        $document = UchebaPrepodavatelyamMakety::findOrFail($id);
         
-        return view('dashboard.pedagogicheskij-sostav-dokumenty-edit', compact('document'));
+        return view('dashboard.ucheba-prepodavatelyam-makety-edit', compact('document'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id): RedirectResponse
+    public function update(Request $request, string $id)
     {
         $validated = $request->validate([
             'title' => 'required|min:4|max:250',
@@ -104,16 +104,16 @@ class PedagogicheskijSostavDokumentyController extends Controller
             ],
         ]);
 
-        $document = PedagogicheskijSostavDokumenty::findOrFail($id);
+        $document = UchebaPrepodavatelyamMakety::findOrFail($id);
 
         // Документ
-        $file = (new \App\Services\FileUpdate($document, 'pedagogicheskij-sostav-dokumenty', $validated))->file_update();
+        $file = (new \App\Services\FileUpdate($document, 'prepodavatelyam-makety', $validated))->file_update();
 
         // Подпись
-        $sig_file = (new \App\Services\FileUpdate($document, 'pedagogicheskij-sostav-dokumenty', $validated))->sig_update();
+        $sig_file = (new \App\Services\FileUpdate($document, 'prepodavatelyam-makety', $validated))->sig_update();
 
         // Ключ
-        $key_file = (new \App\Services\FileUpdate($document, 'pedagogicheskij-sostav-dokumenty', $validated))->key_update();
+        $key_file = (new \App\Services\FileUpdate($document, 'prepodavatelyam-makety', $validated))->key_update();
 
         // Тип файла
         $filetype = array_key_exists("input-main-file", $validated) ? $validated["input-main-file"]->getClientOriginalExtension() : $document->filetype;
@@ -135,7 +135,7 @@ class PedagogicheskijSostavDokumentyController extends Controller
      */
     public function destroy(string $id): RedirectResponse
     {
-        $document = PedagogicheskijSostavDokumenty::findOrFail($id);
+        $document = UchebaPrepodavatelyamMakety::findOrFail($id);
 
         // Удаление документа
         if (Storage::exists($document->file)) {
@@ -143,13 +143,17 @@ class PedagogicheskijSostavDokumentyController extends Controller
         }
 
         // Удаление подписи
-        if (Storage::exists($document->sig_file)) {
-            Storage::delete($document->sig_file);
+        if ($document->sig_file) {
+            if (Storage::exists($document->sig_file)) {
+                Storage::delete($document->sig_file);
+            }
         }
 
         // Удаление ключа
-        if (Storage::exists($document->key_file)) {
-            Storage::delete($document->key_file);
+        if ($document->key_file) {
+            if (Storage::exists($document->key_file)) {
+                Storage::delete($document->key_file);
+            }
         }
 
         // Удаление модели
